Model_deployment/web/requirements.txt
Flask
pymongo
flask-wtf
flask-bcrypt
flask-login
pillow
flask-mail
numpy
pandas
scikit-learn
tensorflow
